Test.MuCheck.MuOp

Description

Mutation operators

Documentation

data MuOp

MuOp constructor used to specify mutation transformation

class Mutable a where

Mutation operation representing translation from one fn to another fn.

Methods

(==>) :: a -> a -> MuOp

(==>*) :: Mutable a => a -> [a] -> [MuOp]

The function ==>* pairs up the given element with all elements of the second list, and applies ==> on them.

(*==>*) :: Mutable a => [a] -> [a] -> [MuOp]

The function *==>* pairs up all elements of first list with all elements of second list and applies ==> between them.

(~~>) :: (MonadPlus m, Eq a) => a -> a -> a -> m a

The function ~~> accepts two values, and returns a function that if given a value equal to first, returns second we handle x ~~> x separately

mkMpMuOp :: (MonadPlus m, Typeable a) => MuOp -> a -> m a

A wrapper over mkMp

same :: MuOp -> Bool

The function same applies on a MuOP determining if transformation is between same values.

type Module_ = Module SrcSpanInfo

SrcSpanInfo wrapper

type Name_ = Name SrcSpanInfo

SrcSpanInfo wrapper

type QName_ = QName SrcSpanInfo

SrcSpanInfo wrapper

type QOp_ = QOp SrcSpanInfo

SrcSpanInfo wrapper

type Exp_ = Exp SrcSpanInfo

SrcSpanInfo wrapper

type Decl_ = Decl SrcSpanInfo

SrcSpanInfo wrapper

type Literal_ = Literal SrcSpanInfo

SrcSpanInfo wrapper

type GuardedRhs_ = GuardedRhs SrcSpanInfo

SrcSpanInfo wrapper

type Stmt_ = Stmt SrcSpanInfo

SrcSpanInfo wrapper

type Alt_ = Alt SrcSpanInfo

SrcSpanInfo wrapper

type Rhs_ = Rhs SrcSpanInfo

SrcSpanInfo wrapper

type Annotation_ = Annotation SrcSpanInfo

SrcSpanInfo wrapper

getSpan :: MuOp -> (Int, Int, Int, Int)

getSpan retrieve the span as a tuple