Test.MuCheck.TestAdapter

Description

Module for adapting test framekworks

Documentation

data Summarizable a => InterpreterOutput a

Wrapper for interpreter output

Constructors

Io

Fields

_io :: Either InterpreterError a

_ioLog :: String

data Mutant

Holding mutant information

Constructors

Mutant

Fields

_mutant :: String

_mtype :: MuVar

_mspan :: Span

toMutant :: (MuVar, Span, String) -> Mutant

Convert a tuple to a mutant

type TestStr = String

Holding test information

newtype Summary

Summary of test run

Constructors

Summary String

class Typeable s => Summarizable s where

Interface to be implemented by a test framework

Minimal complete definition

testSummary, isSuccess

Methods

testSummary :: Mutant -> TestStr -> InterpreterOutput s -> Summary

Summary of test suite on a single mutant

isSuccess :: s -> Bool

Was the test run a success

isFailure :: s -> Bool

Was the test run a failure

isOther :: s -> Bool

Was the test run neither (gaveup/timedout)

class TRun a s | a -> s where

Specify the kind of test suite to run. The result processing, and also how the tests are run are based on this.

Methods

genTest :: a -> String -> TestStr

Generate a runnable test string out of passed in module name and annotated test function name.

getName :: a -> String

toRun :: String -> a

summarize_ :: a -> Mutant -> TestStr -> InterpreterOutput s -> Summary

Functions to be used to summarize

success_ :: a -> s -> Bool

failure_ :: a -> s -> Bool

other_ :: a -> s -> Bool